Fix division in getLoadAverage: cast numSamples to double for accurate average calculation

+++ src/platform/windows/windows_system_monitor.cpp
@@ -247,7 +247,7 @@ private:
             sum += value;
         }
         
-        return sum / numSamples;
+        return sum / static_cast<double>(numSamples);
     }
     
     /**
